+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:soccer_app_flutter/pages/menu_page.dart';
 import 'package:soccer_app_flutter/pages/note_page.dart';
 import 'package:soccer_app_flutter/theme/button_theme_extension.dart';
 import 'package:soccer_app_flutter/utils/layout_helpers.dart';
@@ -73,24 +74,6 @@ class MainNavigationPageState extends State<MainNavigationPage> {
         ],
         currentIndex: _selectedIndex, // 現在のインデックスを設定
         onTap: onItemTapped, // アイテムがタップされたときの処理
-      ),
-    );
-  }
-}
-
-// メニューページ（空ページ）
-class MenuPage extends StatelessWidget {
-  const MenuPage({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        title: const Text('メニュー デモページ'),
-      ),
-      body: const Center(
-        child: Text('練習メニューの内容を記載予定', style: TextStyle(fontSize: 24)),
       ),
     );
   }
